feat: member 사용자정의 repository추가
