+++ src/main/java/com/example/querydsl/domain/yhcourse/controller/MemberController.java
@@ -7,17 +7,18 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.example.querydsl.domain.yhcourse.dto.MemberSearchCondition;
 import com.example.querydsl.domain.yhcourse.dto.MemberTeamDto;
-import com.example.querydsl.domain.yhcourse.repository.MemberJpaRepository;
+import com.example.querydsl.domain.yhcourse.repository.MemberRepository;
 
 import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
 public class MemberController {
-	private final MemberJpaRepository memberJpaRepository;
+	// private final MemberJpaRepository memberJpaRepository;
+	private final MemberRepository memberRepository;
 	@GetMapping("/v1/members")
 	public List<MemberTeamDto> searchMemberV1(MemberSearchCondition condition)
 	{
-	return memberJpaRepository.search(condition);
+	return memberRepository.search(condition);
 	}
 }
