+++ src/main/java/com/example/querydsl/domain/yhcourse/repository/MemberRepositoryCustom.java
@@ -1,0 +1,10 @@
+package com.example.querydsl.domain.yhcourse.repository;
+
+import java.util.List;
+
+import com.example.querydsl.domain.yhcourse.dto.MemberSearchCondition;
+import com.example.querydsl.domain.yhcourse.dto.MemberTeamDto;
+
+public interface MemberRepositoryCustom{
+	List<MemberTeamDto> search(MemberSearchCondition condition);
+}
